# backend/src/main/java/com/example/api_tools/model/User.java
package com.example.api_tools.model;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Public user profile")
public class User {
    private Long id;
    private String name;
    private String email;
    private Integer age;
}